Remove Config.toml from root

+++ src/inputs/Inputs.h
@@ -18,7 +18,7 @@ namespace Inputs
     /*
      * Moves the mouse cursor by the specified amount.
      */
-    void MoveMouseRaw(float dx, float dy);
+    void MoveMouseRaw(int dx, int dy);
 
     /*
      * Returns true if a controller is connected.
